Edits to formatting of code. Null Reference Exception Ball.SettingDegrees().
