+++ Assets/Scripts/Bricks.cs
@@ -7,60 +7,60 @@
 
 public class Bricks : MonoBehaviour {
 
-    private GameControl control;
-    private int lives;
+  private GameManager manager;
+  private int lives;
     
-    /// <summary>
-    /// Call on first frame
-    /// </summary>
-    private void Awake()
-    {
-        SwitchColor();
-    }
-
-    /// <summary>
-    /// Deals with collisions, ball with bricks
-    /// </summary>
-    /// <param name="collision"></param>
-    private void OnCollisionEnter2D(Collision2D collision)
-    {
-        //If the ball collides with the brick game object; Brick looses one life; check to see if you have won
-        if (collision.gameObject.CompareTag("Ball"))
-        {
-            // lives--;
-            // SwitchColor(); // SUBJECT TO CHANGE; SPRITE IS CHANGING NOT COLOR
-            // control.CheckWinCondition();
-        }
-    }
-
+  /// <summary>
+  /// Call on first frame
+  /// </summary>
+  private void Awake()
+  {
+    SwitchColor();
+  }
 
   /// <summary>
-  /// SUBJECT TO CHANGE; SPRITE IS CHANGING NOT COLOR
+  /// Deals with collisions, ball with bricks
   /// </summary>
-  private void SwitchColor()
+  /// <param name="collision"></param>
+  private void OnCollisionEnter2D(Collision2D collision)
   {
-    switch (lives)
+    //If the ball collides with the brick game object; Brick looses one life; check to see if you have won
+    if (collision.gameObject.CompareTag("Ball"))
     {
-      case 5:
-        gameObject.GetComponent<Renderer>().material.color = Color.red;
-        break;
-      case 4:
-        gameObject.GetComponent<Renderer>().material.color = Color.magenta;
-        break;
-      case 3:
-        gameObject.GetComponent<Renderer>().material.color = Color.yellow;
-        break;
-      case 2:
-        gameObject.GetComponent<Renderer>().material.color = Color.blue;
-        break;
-      case 1:
-        gameObject.GetComponent<Renderer>().material.color = Color.white;
-        break;
-      case 0:
-        gameObject.SetActive(false);
-        break;
-      default:
-        break;
+      lives--;
+      SwitchColor(); // SUBJECT TO CHANGE; SPRITE IS CHANGING NOT COLOR
+      manager.CheckWinCondition();
     }
   }
+
+
+/// <summary>
+/// SUBJECT TO CHANGE; SPRITE IS CHANGING NOT COLOR
+/// </summary>
+private void SwitchColor()
+{
+  switch (lives)
+  {
+    case 5:
+      gameObject.GetComponent<Renderer>().material.color = Color.red;
+      break;
+    case 4:
+      gameObject.GetComponent<Renderer>().material.color = Color.magenta;
+      break;
+    case 3:
+      gameObject.GetComponent<Renderer>().material.color = Color.yellow;
+      break;
+    case 2:
+      gameObject.GetComponent<Renderer>().material.color = Color.blue;
+      break;
+    case 1:
+      gameObject.GetComponent<Renderer>().material.color = Color.white;
+      break;
+    case 0:
+      gameObject.SetActive(false);
+      break;
+    default:
+      break;
+  }
+}
 }
